# Conflicts:
#	src/app/modules/legacy/components/cards/object/image/image.html
#	src/app/modules/legacy/components/cards/object/image/image.ts
#	src/app/modules/legacy/components/cards/object/video/video.html
#	src/app/modules/legacy/components/cards/object/video/video.ts

+++ src/app/modules/channels/channel.component.ts
@@ -1,10 +1,10 @@
-import { Component, ViewChild } from '@angular/core';
+import { Component, ViewChild, SkipSelf, Injector, Input } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 
 import { Subscription } from 'rxjs';
 
 import { Client, Upload } from '../../services/api';
-import { MindsTitle } from '../../services/ux/title';
+
 import { Session } from '../../services/session';
 import { ScrollService } from '../../services/ux/scroll';
 import { RecentService } from '../../services/ux/recent';
@@ -12,26 +12,30 @@ import { RecentService } from '../../services/ux/recent';
 import { MindsUser } from '../../interfaces/entities';
 import { MindsChannelResponse } from '../../interfaces/responses';
 import { ContextService } from '../../services/context.service';
-import { FeaturesService } from "../../services/features.service";
+import { FeaturesService } from '../../services/features.service';
 import { Observable } from 'rxjs';
-import { DialogService } from  '../../common/services/confirm-leave-dialog.service'
-import { BlockListService } from "../../common/services/block-list.service";
+import { DialogService } from '../../common/services/confirm-leave-dialog.service';
+import { BlockListService } from '../../common/services/block-list.service';
 import { ChannelSortedComponent } from './sorted/sorted.component';
+import { ClientMetaService } from '../../common/services/client-meta.service';
+import {
+  MetaService,
+  MIN_METRIC_FOR_ROBOTS,
+} from '../../common/services/meta.service';
+import { ConfigsService } from '../../common/services/configs.service';
 
 @Component({
-  moduleId: module.id,
   selector: 'm-channel',
-  templateUrl: 'channel.component.html'
+  templateUrl: 'channel.component.html',
+  providers: [ClientMetaService],
 })
-
 export class ChannelComponent {
-
-  minds = window.Minds;
+  readonly cdnAssetsUrl: string;
   filter: any = 'feed';
   isLocked: boolean = false;
 
   username: string;
-  user: MindsUser;
+  @Input() user: MindsUser;
   offset: string = '';
   moreData: boolean = true;
   inProgress: boolean = false;
@@ -48,7 +52,7 @@ export class ChannelComponent {
     public client: Client,
     public upload: Upload,
     public router: Router,
-    public title: MindsTitle,
+    public metaService: MetaService,
     public scroll: ScrollService,
     public features: FeaturesService,
     private route: ActivatedRoute,
@@ -56,10 +60,23 @@ export class ChannelComponent {
     private context: ContextService,
     private dialogService: DialogService,
     private blockListService: BlockListService,
-  ) { }
+    private clientMetaService: ClientMetaService,
+    private configs: ConfigsService,
+    @SkipSelf() injector: Injector
+  ) {
+    this.clientMetaService
+      .inherit(injector)
+      .setSource('single')
+      .setMedium('single');
+    this.cdnAssetsUrl = configs.get('cdn_assets_url');
+  }
 
   ngOnInit() {
-    this.title.setTitle('Channel');
+    this.updateMeta();
+    if (this.user) {
+      this.clientMetaService.recordView(this.user);
+    }
+
     this.context.set('activity');
     this.onScroll();
 
@@ -70,7 +87,8 @@ export class ChannelComponent {
       this.editing = false;
 
       if (params['username']) {
-        this.changed = this.username !== params['username'];
+        const username = this.user ? this.user.username : this.username;
+        this.changed = username !== params['username'];
         this.username = params['username'];
 
         feedChanged = true;
@@ -82,6 +100,8 @@ export class ChannelComponent {
         } else {
           this.filter = params['filter'];
         }
+      } else {
+        this.filter = 'feed';
       }
 
       if (params['editToggle']) {
@@ -90,8 +110,9 @@ export class ChannelComponent {
 
       if (this.changed) {
         this.load();
+        console.log('reloading channel...');
       } else if (feedChanged) {
-        console.log('reload feed with new settings')
+        console.log('reload feed with new settings');
       }
     });
   }
@@ -100,34 +121,80 @@ export class ChannelComponent {
     this.paramsSubscription.unsubscribe();
   }
 
+  ngAfterViewInit() {
+    this.updateMeta();
+  }
+
+  private updateMeta(): void {
+    if (this.user) {
+      const url = `/${this.user.username.toLowerCase()}`;
+      this.metaService
+        .setTitle(`${this.user.name} (@${this.user.username})`)
+        .setDescription(
+          this.user.briefdescription || `Subscribe to @${this.user.username}`
+        )
+        .setOgUrl(url)
+        .setCanonicalUrl(url)
+        .setOgImage(this.user.avatar_url.master, {
+          width: 2000,
+          height: 1000,
+        })
+        .setRobots(
+          this.user['subscribers_count'] < MIN_METRIC_FOR_ROBOTS
+            ? 'noindex'
+            : 'all'
+        );
+      if (this.user.is_mature || this.user.nsfw.length) {
+        this.metaService.setNsfw(true);
+      }
+    } else if (this.username) {
+      this.metaService.setTitle(this.username);
+    } else {
+      this.metaService.setTitle('Channel');
+    }
+  }
+
   load() {
     this.error = '';
 
     this.user = null;
-    this.title.setTitle(this.username);
+    this.updateMeta();
 
-    this.client.get('api/v1/channel/' + this.username, {})
+    this.client
+      .get('api/v1/channel/' + this.username, {})
       .then((data: MindsChannelResponse) => {
         if (data.status !== 'success') {
           this.error = data.message;
           return false;
         }
         this.user = data.channel;
-        if (!(this.session.getLoggedInUser() && this.session.getLoggedInUser().guid === this.user.guid)) {
+        if (
+          !(
+            this.session.getLoggedInUser() &&
+            this.session.getLoggedInUser().guid === this.user.guid
+          )
+        ) {
           this.editing = false;
         }
-        this.title.setTitle(`${this.user.name} (@${this.user.username})`);
+        this.updateMeta();
 
-        this.context.set('activity', { label: `@${this.user.username} posts`, nameLabel: `@${this.user.username}`, id: this.user.guid });
+        this.context.set('activity', {
+          label: `@${this.user.username} posts`,
+          nameLabel: `@${this.user.username}`,
+          id: this.user.guid,
+        });
         if (this.session.getLoggedInUser()) {
           this.addRecent();
         }
+
+        // this.load() is only called if this.user was not previously set
+        this.clientMetaService.recordView(this.user);
       })
-      .catch((e) => {
+      .catch(e => {
         if (e.status === 0) {
           this.error = 'Sorry, there was a timeout error.';
         } else {
-          this.error = 'Sorry, the channel couldn\'t be found';
+          this.error = "Sorry, the channel couldn't be found";
           console.log('couldnt load channel', e);
         }
       });
@@ -138,7 +205,10 @@ export class ChannelComponent {
   }
 
   shouldShowFeeds() {
-    return ['feed', 'images', 'videos', 'blogs'].indexOf(this.filter.toLowerCase()) > -1;
+    return (
+      ['feed', 'images', 'videos', 'blogs'].indexOf(this.filter.toLowerCase()) >
+      -1
+    );
   }
 
   getFeedType() {
@@ -166,22 +236,19 @@ export class ChannelComponent {
   }
 
   onScroll() {
-    var listen = this.scroll.listen((view) => {
-      if (view.top > 250)
-        this.isLocked = true;
-      if (view.top < 250)
-        this.isLocked = false;
+    var listen = this.scroll.listen(view => {
+      if (view.top > 250) this.isLocked = true;
+      if (view.top < 250) this.isLocked = false;
     });
   }
 
   updateCarousels(value: any) {
-    if (!value.length)
-      return;
+    if (!value.length) return;
     for (var banner of value) {
       var options: any = { top: banner.top };
-      if (banner.guid)
-        options.guid = banner.guid;
-      this.upload.post('api/v1/channel/carousel', [banner.file], options)
+      if (banner.guid) options.guid = banner.guid;
+      this.upload
+        .post('api/v1/channel/carousel', [banner.file], options)
         .then((response: any) => {
           response.index = banner.index;
           if (!this.user.carousels) {
@@ -190,28 +257,27 @@ export class ChannelComponent {
           this.user.carousels[banner.index] = response.carousel;
         });
     }
-
   }
 
   removeCarousel(value: any) {
-    if (value.guid)
-      this.client.delete('api/v1/channel/carousel/' + value.guid);
+    if (value.guid) this.client.delete('api/v1/channel/carousel/' + value.guid);
   }
 
-  async update() {    
+  async update() {
     await this.client.post('api/v1/channel/info', this.user);
-   
+
     this.editing = false;
   }
 
   unBlock() {
     this.user.blocked = false;
-    this.client.delete('api/v1/block/' + this.user.guid, {})
+    this.client
+      .delete('api/v1/block/' + this.user.guid, {})
       .then((response: any) => {
         this.user.blocked = false;
         this.blockListService.remove(`${this.user.guid}`);
       })
-      .catch((e) => {
+      .catch(e => {
         this.user.blocked = true;
       });
   }
@@ -222,18 +288,18 @@ export class ChannelComponent {
     }
 
     this.recent
-      .store('recent', this.user, (entry) => entry.guid == this.user.guid)
+      .store('recent', this.user, entry => entry.guid == this.user.guid)
       .splice('recent', 50);
   }
 
   /**
-    * canDeactivate() 
-    * Determines whether a page can be deactivated.
-    * In this instance, a confirmation is needed  from the user 
-    * when requesting a new page if editing === true
-    *   
-    * @returns { Observable<boolean> | boolean }
-    */
+   * canDeactivate()
+   * Determines whether a page can be deactivated.
+   * In this instance, a confirmation is needed  from the user
+   * when requesting a new page if editing === true
+   *
+   * @returns { Observable<boolean> | boolean }
+   */
   canDeactivate(): Observable<boolean> | boolean {
     if (this.feed && this.feed.canDeactivate && !this.feed.canDeactivate()) {
       return false;
